Deck picker screen with updated due and new counts
Non-cached version of the counts for new and due cards for each deck
listed in the deck picker. Ported quite a bit of python code from
libanki to make it happen (ie. there are probably bugs).

tagIds() in Deck.java is only declared, but
not implemented since it will not be called from the current code. However
it needs implementation for possible future calls to it.