Change order of SRS operations
Changed back to the original ordering of the SRS operations, i.e. first
answer the card and then get the new card. No more suspending of the
current card while getting the new card, and then answering it. Since
adding transactions, testing has shown that this order is actually more
responsive. It is also easier to maintain since it is the same as
desktop Anki.