Allocate the payload data as an Object array.
In case of failure downloading media, we were putting in the first
element of the array the name of the file. However, since the array was
allocated as an array of strings, this was causing a runtime exception.

Instead, allocate an array of object (the objects in it are still
integers) and allocate a new one-element array for the error case,
because that is what the caller expects.