Ignore next and prev when disabled.
Since we cannot remove a pending intent associated with the prev and next arrow, associate an action
that is ignored by the service instead.

This prevents the user from being able to press next and go beyond the last deck.
Moving backward from the first item was also possible, but was leading to the first item itself, so it
was not a visible issue.