Add support for custom fonts.
This change introduces a way to load custom fonts in the WebView showing
the card.

The custom fonts must be stored in the "fonts" directory under the deck
path (by default, "/sdcard/AnkiDroid/fonts/"). When the card is
generated, a set of CSS @font-face directives are generated for each of
the files found in the directory. The font-family associated with each
fonts is the name of the font file, without extension.

For example, if your card template uses the font-family "verdana" and
you want to use your own font to render this on the Android device in
Anki, you should copy a file called verdana.ttf to
/sdcard/AnkiDroid/fonts/verdata.ttf. This can be done either using the
command "adb push" or by mounting the SD card on a computer and copying
the files to the directory /AnkiDroid/fonts/ in the SD card.

Note, that any file found in that directory will be added as a
@font-face directive, even if the file is not a font, or the font format
is not recognized by Android.

Fix issue 565: Use mincho font.