Fetch deck status asynchronously.
The deck status is now stored in the MetaDB and computed using an
AsyncTask (therefore not in the main thread as it used to be).

This AsyncTask is started when an update is necessary and it triggers
the AnkiDroidWidget.UpdateService when it is done.

The deck status is stored in the MetaDB.