Avoid failure when closing the currently open deck.
Currently all decks on the same file share the same database. This
assumes that there is only one deck for each file. However, the async
task that handles fetching the details of a deck for the widget
introduces a second deck open on the same database. Closing the database
that is used by someone else causes an issue.

The previous solution was to close the database and re-open it after
scanning all the decks, however, this is no longer sufficient because
the scan now happens in a background thread and therefore possibly
concurrently with other operations on the database.

Instead, close the database only if it is not the current deck in the
application. This is not (yet) 100% bullet proof because it is still
possible for someone to open the database between the time it is
compared with the current deck and the time it is actually closed.