Fixes color of buttons text on Xoom.
The Xoom uses a different theme, where the buttons are semi-transparent
and the text color is white. This works well on the activities with a
dark background (like the study options activity) but not in others
(like the flashcard review activity) it is not readable.

This commit introduces a style to be applied to the buttons that is
empty by default but that defines the text color to be black for the SDK
version 11 (Honeycomb). This is achived by creating a values-v11
directory and redefining the style in a file named styles.xml in that
directory.

For the deck picker activity instead, remove the gray bar under the
button at the screen: it does not seem necessary and it does not match
the content of the list above it. This prevents the problem with the
background and the text color being too close (light gray and white).

Issue 609: Buttons at the bottom of the card are hard to read on Xoom