Use two different WebView instances to show the card.
The main reason for this is that there is a leak in WebView related to
loading fonts. When a font is loaded a large amount of native memory is
allocated, and re-allocated each time a webpage is loaded. This quickly
leads to the process being killed because it uses too much memory.

The solution applied here is to release the WebView each time a new page
is loaded, which releases the memory used to load the font in the first
WebView.

This addresses the following:
http://code.google.com/p/ankidroid/issues/detail?id=687