Fixes themes for Honeycomb.
The application on Honeycomb needs to derive from the Holo and
Holo.Light themes in order to have the action bar, which is the only way
to access a number of the functionalities, including the menu.

This commit adds a copy of themes.xml in the values-v11 directory, where
the themes used by the application derive from the Honeycomb base
themes.