Fix for issue 618: Decks with Japanese names.
Issue 618: Ankidroid does not display deck titles with Japanese names
and fails to sync said decks.

As it turns out, the behavior of Anki on the desktop is different for
shared and personal decks.

Shared decks are limited to 40 characters and remove all but basic
characters from the file name.

Personal decks, on the other hand, are not stripped at all, to make sure
they can be synced correctly.

This makes AnkiDroid more consistent with Anki on the desktop, and
should make both shared and personal decks download work as expected.