Make sure the media directory exists.
Upon completing the download of a shared deck, if there is any media,
the media should be copied to the .media directory.

However, the directory has to exist.

There used to be a statement creating the directory, but it was part of
a logging statement and it was accidentally removed. This means media
for shared deck are currently broken.