Add a settings to disable WebView switching.
WebView switching was added in Android 0.7 to fix a bug with custom
fonts on some versions of Android. However, recent versions of Android
seem to have fixed this issue.

This commit adds a setting to force quick updating of WebViews, so that
we can ask users to test this on a number of devices with the next
release and see if this is still an issue on any device.

Hopefully, we can identify which versions of Android we can disable the
workaround by default and possibly remove the setting in future
versions.