Added evernote app integration
Requires the Evernote Android App!
Creates an anki note for every evernote note matching a tag specified in
the general preferences. Doesn't synchronize contents but only the
Evernote note ID. While reviewing the anki notes the associated Evernote
note will be opened in the Evernote App. Tap the return key to get back
to AnkiDroid and rate the difficulty.

!!! VERY EARLY STATE !!!