Do not use onCancelled with result.
The overloading that takes a result has been introduced in API Level 11
and therefore this causes a crash on previous version.

The default implementation is to simply call the version that does not
have a result, so we can use that instead. None of our code should
depend on the former version anyway, since we cannot guarantee it will
be called.

Issue: https://code.google.com/p/ankidroid/issues/detail?id=1769