New night mode icons.
This commit adds two new icons for night mode on/off status. The icons
represent the moon and the sun, and they are meant to indicate the state
the user will transition to, i.e., the moon icon is used to enable night
more and the sun icon is used to disable night mode.

Also add the original files at 320x320 resolution the icons were created
from, as well as a simple script to generate the assets at the various
resolutions.

The icons were created by Flavio Lerda (flerda@gmail.com) and are
released under the same copyright as the AnkiDroid source code.