Move SDK specific methods into Compat.
On certain devices (e.g., Donut), the application crashes if a class is
using a method that does not exist with a VerifyError even before any
methods on the class can be invoked.

To avoid this, we introduced a Compat interface, which hides all the
platform specific method calls. However, the disableWriteAheadLogging()
method was being called on SQLiteDatabase without following this
pattern, leading to a crash on start-up.

This commit moves those calls behind the interface.