Show terms and conditions with shared decks.
Currently the shared deck screen does not include the terms and
conditions, which are actually useful on that screen, as they cover the
use of shared decks.

Add them at the bottom of the screen so that they are always visible.
Clicking on them will open the default browser and to show the terms.