Pre-select current deck in card editor.
Make add card behavior similar to Anki: when a user adds a card from the
study options or the card browser, pre-select the current deck.
AnkiDroid does not have the notion a currently selected deck in the deck
browser, so adding a card from there will still default to the Default
deck.