Proposed fix for Issue 1787: Duplicate can't save.
I'm proposing this fix for 1787. When users attempt to save a duplicate
note, they will be notified of the shared primary key and asked if they
wish to save it regardless. Initially I was worried this duplicate
prevention functionality was necessary for a reason unknown to me, but
after browsing through the note update code it looks as though there's
no entry collision issues when updating since it's using the unique
note ID anyway.