Fix for Issue 2068: Missing autoplay in answer.
This is my proposed fix for issue 2068, where audio before the <hr
id="answer"> would not autoplay when the card answer is shown. This was
caused by fix 18fdb7b, which used mCurrentCard.getPureAnswerForReading()
to grab the text after the <hr> to parse for audio. I've changed this to
parse the whole answer card to match Anki Desktop.