shared random number between question and answer sides
(Related to pull request https://github.com/dae/anki/pull/74 for Anki.)

It's unfeasible (and pretty useless) to generate cards for numbers individually. Having a shared random value between the question and the answer (which is impossible otherwise) would make it possible to create cards like "Numbers between 1 and 10" or "Tens: 10, 20, ..., 100" or "Factors of ten: 10, 100, 1000, ..."

A card, with javascript, can generate the number and the corresponding answer in the foreign language. Consider how fun it would be to be able to write something like this:

    <script>
      // get a number and the number of zeroes between the set limits
      var number = Math.floor({{Rand}} * ({{Max}} - {{Min}} + 1) ) + {{Min}};
      var zeroes = Math.floor({{Rand2}} * ({{Zeroes Max}} - {{Zeroes Min}} + 1)) + {{Min}};

      // compute the final number
      number = (number+{{Increment By}})*{{Multiplier}}*Math.pow(10,zeroes-1);

      document.getElementById("question").innerHTML = number;
    </script>