fix issue 1878 - filter and select all tags.
Checkbox for selectall tags and filter tag added.
The filter tag in the cardeditor is used with the edittext already in there to add a tag.
For others dialogs, I had to add an edittext field to the layout.

RadioButtons in the custom study dialog based on tags with forced white_btn_radio style.