Disable media sync on Android <=2.2 if non-ascii file name found
Unicode normalization is not supported on Android 2.2 and below. Since the new media sync protocol requires file names to be normalized in NFC form, we now abandon a media sync if we detect a media file name that cannot be entirely represented in the ASCII encoding and disable media syncing on that device.

This approach is likely to allow a number of users on those devices to still be able to sync media without issue.

Note: A dialog is shown if media becomes disabled after a sync. Collection syncing is unaffected.