Fixed Media Sync file not found error
Worked around the temp file not found error by using ZipInputStream
(does not require the file to be written to the disk). In any case, the
data was already in memory from the HTTP input stream, so no additional
device memory should be used than before.

Finished Testing

A slight increase in memory usage, on account of that the meta data is
fetched first. Probably it is the same as any normal copy spike.

Undid temp change

Restored empty line

Removed commented code

Finished Testing

A slight increase in memory usage, on account of that the meta data is
fetched first. Probably it is the same as any normal copy spike.

Undid temp change

Restored empty line

Removed commented code