Use CSS zoom to scale card content
The previous approach relied on a complex search for HTML elements
containing text and the CSS rules to style them. It is much simpler and
more efficient to zoom the entire card (including images) with a single
CSS rule.

The existing card and image scaling preferences were renamed to better
reflect their purpose and moved to a more sensible section of the
preferences screen. This commit will also update the old preferences
on the next version upgrade.