Import rework
Rewrote the structure of the importers to match the original.

Updated the Anki 2 .apkg importer to support all features of the desktop
client (i.e., updates and dupe handling). Improved error handling and
progress updating.

Collection objects must now be created with a Context. The Context can
be retrieved wherever a Collection object exists. As a consequence of
this change, some open/close logic in the full sync code has been
reviewed and slightly reworked.