PoC of forecast statistics for month view using greengreen's algorithm.
Get the gregreen algorithm for forecast statistics working in AnkiDroid
for the month view.

TODO:
- Remove the many transformations of the data by using the
representation required for plotting in the algorithm. Now it does
matrix transpose to go from the representation of the algorithm to
representation of the dues matrix. Then it goes from dues matrix to
representation required for plotting. Then it cumulates. Also re-use the
Card object and work with a cursor instead of loading all Cards in
memory since we do one Card at a time.
- Remove hard coded things which really should not be hard coded (like
deck creation timestamp and now()).
- Use CamelCase in code.
- Predict 30 days, not 29.
- Remove sysouts