Support multiple decks, Use per-deck settings (max #new cards and #revs per day), do not pre-load all cards
Notes:
- There seems to be a bug in AnkiDroid which overwrites the max. # of
new cards for all decks if it is updated for 1 deck.
- Using an iterator with an open cursor instead of pre-loading all cards
seems to make it slower. Maybe revert at a later time.
- I used to pass mCol from stats.java to the hook and replaced it with
context so that AdvancedStatistics could get mCol from the
CollectionHelper. Maybe revert that as well. I did that because I
thought that the problem in note (1) occurred in the statistics because
I didn't reload the settings after they where changed. I thought
reloading using the collectionhelper would fix this. But it didn't.
Since the problem also occurs in the deck picker it seems safe to revert
this and to look for the problem in AnkiDroid itself.