Enforce correct max # reviews per day, get collection.crt from collection instead of hardcoded, add javadoc, cleanup code a bit
Enforce correct max # reviews per day by keeping track of all reviews
per day instead of per timeBin. I used to check that the # reviews per
timeBin shouldn't exceed the max # reviews per timeBin (= max # reviews
per day * days per timeBin). That is not correct if on a particular day
within a timeBin the max # of reviews per day is exceeded, but the max #
of reviews per timeBin is not. It also would allow the 'first' cards in
a timeBin to be reviewed and the cards which were scheduled later in
that time bin would not be. Using the max # reviews per day limit they
will. But on the other hand we review per card so the first cards
encountered will already fill up future bins in the beginning of the
simulation. Still keeping track per day is better...