Add share question button
I added a button to the toolbar menu that lets the user share the question with other apps. This is useful to look up additional information about a particular card. Most dictionary apps can recive text content to search for data.

This button takes the "question" from the first item of the fields array of the cards note, and then sends an intent with that data to share text content.