drawArc/fillArc: Avoid drawing full arcs when arcAngle is 0
This fixes issues with the card types graph showing almost everything as
black when unseen is 0, except the yellow part for suspended which is
drawn afterwards.

The documentation for Canvas.drawArc says:

    If the sweep angle is >= 360, then the oval is drawn completely.
    Note that this differs slightly from SkPath::arcTo, which treats the
    sweep angle modulo 360. If the sweep angle is negative, the sweep
    angle is treated as sweep angle modulo 360

There's no explicit mention of what happens if the sweep angle is 0,
but it looks like in practice it's the same as if it was 360.

Since the operation is going to be a no-op either way, it doesn't hurt
to skip it.