Add multitouch capabilities to whiteboard for interfacing a card under it.
This makes the whiteboard more ergonomic (in my subjective opinion).
Related issue: https://github.com/ankidroid/Anki-Android/issues/4522
Summing up, when keeping one finger on the whiteboard while it is activated, it is now possible to use a second finger to scroll the card behind the whiteboard or tap onto one of its elements.