#1463 Recognize TTS elements in questions and answers. (#4677)
If the question or answer contains any <tts> elements with the 'service'
attribute set to 'android', then only the text enclosed in these elements
is read aloud using the text-to-speech engine. The 'voice' attribute can
be used to specify the language of the text (details below). If no
<tts service='android'> elements are present in the question or answer,
the user is asked to select the language from a list (as before). <tts>
elements with other values of the 'service' attribute are ignored, which
makes it possible to use the same card format with AnkiDroid and with
the AwesomeTTS add-on to the desktop Anki application (each ignoring
unrecognised services).

'voice' attributes should be given in the format used by the
Locale.toString() function, i.e.

language_code + "_" + country_code [+ "_" + variant_code]

(the second underscore and the variant are optional and typically
unnecessary). Typical examples: en_US, en_GB, fr_FR.

Full example:

<tts service="android" voice="de_DE">{{Back}}</tts> (pl. {{Plural}})

Here only the contents of the 'Back' field are read aloud (in German); the
contents of the 'Plural' field are only displayed, not read.