Keep track of template add/deletes for later persistence
To work well with in-memory template structural changes (adds and deletes)
you have to keep track of them in order, as deletes when persisted alter
card ordinals in the database.

Also, you want to avoid adding a template you will later delete as it
will unnecessarily generate cards and clog up undo functionality.