Add Mathjax support.
This adds Mathjax 2.7.0.  It modifies several files in libanki, but I
tried to check them against the originals in Python.

It does not add fading like Anki but rather shows the text as soon
as Mathjax says it is finished.

I have added a few tests.  Some are ported from Anki, and other are
original.

This particular commit does not actually add the Mathjax files.  These
are in the next commit.