Fix analytics opt in
Even the default in UsageAnalytics class is 'false' the default value in the preferences was set to 'true'. That means analytics was disabled by default but if the user opened the preferences screen it got automatically enabled. This commit fixes this issue by setting the default value in the preferences to 'false'.