Make Reviewer reschedule function closer to Anki Desktop (#5344)
Reschedule works a bit different in AnkiDroid compared to Anki Desktop,
in the sense that Anki Desktop pops up a custom dialog for reschedule,
which let the user choose between rescheduling as a new card (reset progress),
or rescheduling as a review card in x days (or rather between x~y days),
whereas we avoid the custom dialog on AnkiDroid because it's quite a lot of work,
and it deviates even further from material design principles than we already do.

Here I've made a few minor changes that make the implementation of the reschedule
function in the AnkiDroid reviewer a little bit closer to the original implementation
of the reschedule function in the Anki Desktop Browser:

1. Remove the "change position" action. This is a totally separate feature from reschedule
in Anki Desktop, and I argue that it only makes sense in the context of the browser,
with multiple new cards IMO

2. Rename the top level menu from "Schedule" to "Reschedule", and the second level action
to "Reschedule for review in x days". These renames put the feature back inline
with the original Anki Desktop feature, and are more intuitive IMO

3. Move the top level menu from the bottom to just after "Delete note".
Options should always be the last item in the menu to make it easy to access,
and the Reschedule feature logically should be grouped together with the other
card/note related actions.