favour most recent modified when importing template changes
dae/anki@645f6714d2a7fc7e747aeee9f4a7150770e7d8d1

The test must be changed, because when a model is imported, it's "mod"
is set to current time. Which makes it the most recent version of the
model. While actually, it may be hold and should be overriden by the
new version of the model found in the second import.

This is a bug which I was able to reproduce in Anki. Actually, I'm not
even sure it's a bug, it just a strange way to consider things.
Anyway, this allows to test that import deals correctly with models
imported again