Correct assertEquals parameter order
According to
http://junit.sourceforge.net/javadoc/org/junit/Assert.html , in
assertEquals, the expected value should be given before the value
tested. Otherwise, it leads to messages such as
>Error in testAnki2Updates(com.ichi2.anki.tests.libanki.ImportTest): java.lang.AssertionError: expected:<0> but was:<1>
where the expected value is 1 and the value actually tested is 0.

While it certainly does not change any test result, it makes the
debugging process slightly more complex, since the error messages are
misleading.

This is why I did read the 130 assertEquals found by grep, and changed
the order of arguments when it seems to make sens. More precisely, I
assumed that:
* if one value compared is either a literal, or a constant (as L438 of
  ContentProviderTest.java), then it is the expected value
* if two variables are compared, I assume that the one which has been
  set the most recently is the one tested. (For example
  `assertEquals(after, before + 1)`; becomes `assertEquals(before + 1,
  after);` l198 of ImportTest.java. Indeed, `before` was set L177 (and
  even tested L182) while `after` was set L197)
* if a value is named «expected» then it is the expected value. For
  example L156 of ImportTest.java