Correct due of generated siblings.
This commit port current Anki code to Ankidroid. It ensures that when
a note is edited, and the edit generate new card, those new cards
obtain the same due value as their siblings.

The bug was reported (by me) as #5456. And I tested this correction
simply by trying to reproduce the bug with this new code, and checking
that it does not occur anymore.