New cards of existing note take due value of NEW siblings
This commit port to Ankidroid a PR I just did on anki's
library. arthurmilchior/anki@3f45de87b393b14abb0891dc851723d75742aabb.

https://github.com/dae/anki/pull/339

Once again, I tested this by trying to reproduce the bug with this new
version of ankidroid and checking that the bug does not exists
in this version; and that all due numbers were correctly generated.