Remove useless catch of JSONException
This was done the following way:
=====
greping all "catch (JSONException"

If it only did sends a RuntimeException, then I delete it's try, I
delete the catch, and indent again everything between them. (Actually,
an emacs macro did it).

Otherwise I kept it.

If it's in a multiple catch, then I only remove the JSONException
from the multiple catch.

It can be tested as follows
=====
Using gitk, you can ignore change in space. (This is important because
a lot of strings had to be unindented), you can check that the only
things which were removed were:

```java
catch (JSONException x) {
   throw new RuntimeException(x);
}
```
and
```java
try {
```

and some JSONException in multi catches