The word model should be "note type"
Model is used in the code. But the user only see "note type" in anki,
in almost all ankidroid, and in the manual.

Is there a way to ensure that translations are also updated ? The same
problem occurs in French at least. If it can't be done this way,
another solution would be to also create entry saving_note_type toast_last_note_type
rename_note_type and note_type_delete_warning. If possible, tell
ankidroid to use the note_type variant if it exists locally, otherwise
the model variant. After some time, when most translation are up to
date, deleting the "model" entries.