Correcting template to card type
I should note that anki's manual and interface are not consistent with
anki's code and anki's json. In the manual and interface, a note type is
composed of multiple card type. Each card type contains a front and a
back template, which is the HTML displayed, with some mustaches.
In anki code and json, a note type is composed of templates.

For the sake of consistency with what the user see, I renamed
"template" to "card type" when it makes sens. I.e. in a single place.

Same remarks than in last commit applies