New model: more simple and more unique
This commit simultaneously corrects two problems, in which Ankidroid
differs from Anki.

When a standard model is constructed, Anki add random symbols only if
the model's name is already used in the collection. This is now the
case in Ankidroid too.

When a model of name m is cloned, if "m - copy" already exists, then a
random string is appended. This is now the case in Ankidroid too.