factorize modSchema and modSchemaNoCheck
try/catch is slightly expensive, but since it's not executed in loop,
there is no real time gain here.

However, it allows to avoid a
    } catch (ConfirmModSchemaException e) {
        // This will never be reached as we disable confirmation via the "false" argument
        throw new RuntimeException(e);
    }
which is quite ugly and totally useless.

This is simply done by doing the actual work directly in the method
where the exception should never occur, and calling this method from
the method sending the exception.