Improve renameField
This method is called only after that schema is changed. I take this
into account to use renameFieldSchemaModified. It ensure that the
method does not throws ConfirmModSchemaException. Finally it allows to
avoid a try/catch.

In future commit of the same PR, it allows to edit new model without
doing a full sync.