Model's method don't require full sync on new method
This port dae/anki@9863b6f9ff2d75763bcc055395a35a8cb1f43a19
It ensures that in Model, the method requires a new sync only if
necessary. I.e. not on new models.

I should note that this does not ensure Ankidroid does not require
full sync on new model. Indeed, it modifies the collection in the GUI
and not only in Models.java
This will be corrected in future commit