Remove some useless modSchemaNoCheck
In both cases, the methods are only called after a call of
_modSchemaIfRequired. So if model should have been changed, it is.

Without doing this, it's impossible to edit new note type without full
sync.

If we wanted to be even more sure of this code, we could assert that
the model does not have to be changed, using
`isModSchemaRequired`. However, I believe that this would be useless
and only add some code which should have no effect.