use _modSchemaIfRequired in addField/addTemplate
Actually, what we really want to know is whether schema should be
modified or not. Using a single method avoids duplicating function.

And, to be more honest, there are cases in future commmits where we
know that schema does not have to be changed, but we don't know
whether it's because the model is not yet synchronized (and thus can
be changed without full sync), or whether it is because schema is
already marked changed. So it's far easier to test both cases
simultaneously