Adding flag image
Currently black is not used. I'm not sure whether «no flag» is better
represented by a black or a white flag. May be by a flag with a cross
on it, but I don't know how to do such an icon. In doubt, I leave
black flag, so it's easy to edit to use it if one want.