Allow .zip after colpkg/apkg
When downloading an apkg, some browser adds a .zip to the name of the
file (because it's detected as a zip). This leads to .apkg.zip and
.colpkg.zip files.

This commit ensure that in those cases, the file can be imported
without being renamed. This is extremly useful to test other commits,
especially the ones related to the import package.

Since it only accepts .zip if they are preceded by apkg or colpkg, the
risk is relatively small. Even if there is always the risk that
someone tries to zip a file foo.apkg, which becomes foo.apkg, and
should not be read by ankidroid.