Correcting #5518
doInBackgroundUpdateNote is called in a single place, doInBacground,
when the type is TASK_TYPE_UPDATE_FACT. This type is actually used in
two places, both cases in `onActivityResult`. I'll first explain why
CardBrowser can't be the problem.

Card Browser
------------
The task data is created with sCardBrowserCard. It's sent only if the
activity is EDIT_CARD. This activity is sent only once in the code,
just after setting sCardBrowserCard to
getCol().getCard(mCurrentCardId); which is a value which can't be
null.

sCardBrowserCard is not edited anywhere else in the code.
So we have both a garantee that this variable as been set once and
that it's not set to null anywhere else.

AbstractFlashcardReviewer
-------------------------

The task data used to be created with variable `mCurrentCard`. It's
sent only if the activity is EDIT_CURRENT_CARD. This activity is sent
only once, after `sEditorCard` is set to `mCurrentCard`.

It should be noted that `mCurrentCard` is set in multiple places. In
particular in nextCardHandler, where it may becomes null. This would
be the case if for some reason, Ankidroid asks for the next card
before the activity has time to tells the viewer that it did
ends. While it seems hard to reproduce voluntarily, I do imagine it
may occurs on a really slow smartphone.

Note that if I'm right, if `mCurrentCard` is set to another card, it
means that the activity `TASK_TYPE_UPDATE_FACT` will be applied to the
wrong card (which is not really a problem. It means that the correct
updating will simply be delayed until next time there is a db check or
an "empty card" check)

Since `sEditorCard` is not set anywhere else, appart before calling
the activity, there is no more risk in using `sEditorCard` than in
using `mCurrentCard`. In normal usage, they should both have the same
value. In case of trouble because the activity is slow to call
onActivityResult, then this is the variable having the card we care
about.