mCards is not changed while used
I ensure that mCards is set to a new array instead of cleared.
Next aciton is to notify the adapter that the data set has changed, so
anyway the change will be taken into account. However it avoids a
problem if a loop is currently using the array.

More details given in a commit explaining how it relates to #5516