Hopefully correcting #5516
The error occurs here:
https://github.com/ankidroid/Anki-Android/blob/605bad8187d1cf1150eb96306d2fec830d54d409/AnkiDroid/src/main/java/com/ichi2/async/DeckTask.java#L881
Which is extremly strange. The third tests throws an error, stating
that the size of the list is 0. While the second test did assert that
i=446 is less than the size of the list. Which means that the size of
the arrayList was changed between both tests. Since the program is
asynchrone, it's actually a bad idea to change objects during a
run. It's safer to create new object and assign it to the variable. In
any case, there is not even a loss of efficiency for time. And as far
as I understand the code there is nowhere in the code which needs to
see that the list change. Indeed, each time the list is changed, a
method is called to signal the class that the list is changed and that
it should update and do whatever it has to do with new data.

It remained to figure out the entire history of the variable `items`
which we realized could change. The remaining of this commit show you
my long quest to find all of the possible history of this variable

I tried to follow the object `items`.
* It's the first element of `params`
* `params` is an argument passed to DeckTask#doInBackgroundRenderBrowserQA
* this method is called only in DeckTask#doInBackground when type is
  `TASK_TYPE_RENDER_BROWSER_QA`.
* This task is called only in two places. CardBrowser#onScroll and
  CardBrowser#onScrollStateChanged.
* In both cases, the first parameter is simply the result of
  `getCards()`
* `getCards` returns `mCards` (and initialize it if necessary)

so it remains to follow both `mCards` and `getCards` and see when are
elements removed from the list they point to. (Note that assigning
anything to mCards would not create trouble. The only trouble is when
the object itself is modified).

`getCards`
---------
I can safely ignore when the result of getCards is directly used to
get some value. Here, I consider the other uses.

getCards is used once to remove some elements, in
removeNotesView. This is removed in this commit

getCards get cleared in searchCards. This is removed in this commit

It's given as a data to mCheckSelectedCardsHandler
`TASK_TYPE_CHECK_CARD_SELECTION`. So we need to follow this path.

`TASK_TYPE_CHECK_CARD_SELECTION`
-------------------------------
This task is only used in DeckTask#doInBackground, where it calls
`doInBackgroupndCheckCardSelection`, in which the list of cards is
only used to get elements. The list is never changed, so this paths
ends here.

`mCards`
---------

`mCards` is never assigned to anything. It's returned only by
getCards. When it is accessed, it is to get it's element, not to set
them. So no change can occurs directly to it.

mCards can takes values either an empty list, or values[0].getCards();
in `onProgressUpdate` of `mSearchCardsHandler`.
https://github.com/ankidroid/Anki-Android/blob/605bad8187d1cf1150eb96306d2fec830d54d409/AnkiDroid/src/main/java/com/ichi2/anki/CardBrowser.java#L1485

So we now need to follow the list given in argument to this value.

`mSearchCardsHandler.onProgressUpdate`
--------------------
This handler is used only in CardBrowser#searchCards, where it is
passed to DeckTask#launchDeckTask. In it, this listener is given to
create a new DeckTask, where it is assigned to mListener.

`mListener.onProgressUpdate`
----------
mListener.onProgressUpdate is called only in
DeckTask#onProgressUpdate, given it the value it did receive.

`DeckTask.onProgressUpdate`
---------------------------
To follow this methods, if I can trust
https://developer.android.com/reference/android/os/AsyncTask.html#publishProgress(Progress...)
I need actually to follow

`DeckTask.publishProgress`
---------------------------
publisBrogressi s actually called. Let us restrict our search.

The task I need to follow is TASK_TPE_SEARCH_CARDS

`TASK_TPE_SEARCH_CARDS`
----------------------------------------------------
This task may be cancelled. We do not need to consider this. It can be
called in the place we considered above. It is executed in
DeckTask#doInBackground
where it calls DeckTask#doInBackgroundSearchCards.

`publishProgress` is called with the data being `searchResult`, which
itself is the result of Collection#findCardsForCardBrowser. This
method returns a new array which will never be edited after it's
created. So we don't need to investigate Collection.

`searchResult` is accessed and never edited in this method.

This ends our quest and show that normally the list is edited nowhere
else, so that the above-mentionned problem should not occur anymore.