Getting due correct when generating cards with no new siblings
Here is a silly problem with porting Python code to Java. Sometimes,
it does not work the same way at all. When I committed
504d96a99d31f35cfb48d94af1e7b282603b9bb6 I didn't pay attention to was
the value of dues.get(nid) when nid is not in dues. And actually, the
value is None. Which means that it can't be converted to a long. And
thus raised an exception.

I also realized that the line `due = dues.get(nid);` was present twice
in the code, which makes no sens in the first place. So this second
occurrence is now removed.