Correcting #5515 hopefully
The error was sent in this line:
https://github.com/ankidroid/Anki-Android/blob/7235a5bc667a6c9def934d85764cedc6594814b3/AnkiDroid/src/main/java/com/ichi2/async/DeckTask.java#L857
where searchResult is null.
It's value is the result of col.findCardsForCardBrowser
this method simply returns the  value returned by the Finder#findCardsForCardBrowser

The value returned by the finder itself findCardsForCardBrowser sends
null if the task was cancelled. However, the value returned is still
used by findCardsForCardBrowser. So I ensure that it's not null, but
an empty list.