Correct conf' when it's set.
Currently, the conf is corrceted only when accessed. Which I believe
may be the cause of bug #5523. Indeed, data is corrected, and then
accessed, while it is theoretically possible that data is changed back
to a faulty value during an asynchrone process after correction and
before access.

With this change, data is changed BEFORE conf' is saved; which means
that the value is never faulty.

This correction should become useless for people using Anki 2.16 or
latter, thanks to dae/anki#347.