Ending symbol of multiline xml tag on their own line
When a multiline tag is edited, sometime, there are useless edition in
git only because it means that the ending symbol has moved. While in
fact, it could have been on its own line all along. That's why I'm
doing a single commit with > and /> on their own line, so that future
edits are cleaner