Dealing with ":::" in deck name
The method `parent` as I implemented it, does not deal with names
having ":::" in them. "::::" is impossible, but ":::" is totally
okay. So now it deals with it.

Anki believe that in this case, the first two colons are the
separators, the last one is part of the deck name.

As requested by Mike Hardy, I tried now to keep as close as possible
to Anki's code