Set some methods to private
There is actually two reasons to set `removeLrn` and `_revForDeck` to
private.

First reason is that those methods are not called anywhere except in
the class where they are defined. So we can do it without loosing
anything.

The second is that `removeLrn()` and `_revForDeck(long, int)` are not
implemented again in SchedV2. Which means that the type of SchedV2
states that there are public methods `removeLrn()` and
`_revForDeck(long, int)`, while in fact those methods should NOT be
used. Indeed, those inherited methods refere to private variable of
Sched, which are never changed, and ignore the private variables of
SchedV2 with the same name. This would lead to quite non-sensical bug
if for some reason thoses methods were one day used with SchedV2.

Furthermore, once this change is done, it helps create the change
which will occur in the next commit.