Create AbstractSched class
In Python, SchedV2 does not inherits from Sched. There is no reason
for it to be the case in Java. Since every single methods are
implemented again in SchedV2, even if they are identical (as in
Python), this inheritance makes no sens at all.

The only reason for the inheritance to be used here, I believe, is
that it allows to uses the type «Sched» for variable, whatever
scheduler we use.

The correct java way to do this is to have an abstract class or an
interface, inherited by the two classes. So this is what this commit
does.

Actually, the subclass DeckDueTreeNode is used by both Schedulers,
even if it was declared in Sched.java. In order to be consistent, I
did move it to the AbstractSched class.