Add support for Anki 2.1 flag feature (#5571)
* get/set user flag in Cards

* Add method setUserFlag in collection

As in anki

* Adding flag image

Currently black is not used. I'm not sure whether «no flag» is better
represented by a black or a white flag. May be by a flag with a cross
on it, but I don't know how to do such an icon. In doubt, I leave
black flag, so it's easy to edit to use it if one want.

* Add flags' colors in color ressources

* add flags colors in themes

* Add flag colors as colors attr

* Add flags's menu's text in core

* onFlag in AbstractFlashCardViewer

* Add flag in reviewer menu

* Add flag button on reviewer's menu bar

* Change color of flag depending on current card

* Change variable name to "card"

Because `dataSet` or `map` are correct type indication, but gives no meaning

* remove "flags" from Card

Actually, it's value is unrelated to the card's flag; that's
confusing. Instead I use the origanl value of interest: tags and
whether card is suspended

* getColor returns color ressource

The way it was done before was uselessly complex. Creating an ad-hoc
array, with positions as constants. Then returning an index of array.

Now, getColor simply return a ressources. The resource is then used
accessed to access the color.

* add "flags" in cards in browser.

For some reason, instead of using a Card object, cards are encoded as
Map<String, String> and then displayed in the browser. Anyway, to
display correctly, one value was missing: the flag. It's now added in it.

* Use flag to color card's browser.

Now, flag is used. And marked/suspended is used only if the flag is
0 (i.e. no flag)

* Add flag button in browser manager

* add ListenerWithProgressBar

Factorizing plenty of listeners

* Add ListenerWithProgressBarCloseOnFalse

Factorizing even more listeners