Card browser's code improved
Previously, card browser used a special version of the finder, which
returned some "card"s which were in fact Map<String, String>.
There were two problems with this. Firstly, this map needed to
contains all informations which may be potentially useful, such as
question and answer, which may be costly to compute and useless (if
questions/answers are not currently displayed)

Furthermore, this led to code which is more complex than
required. First because card objects should be expected to be either
cids or Card. And not some particular things done for this special
purpose. Especially since Card already contains all required
informations (directly or through note). And it's possible to ask
cards to precompute question/answer and to save the result.

This way, we can remove plenty of methods which were used either to
created those maps, or to update those maps. (And to save those maps
in TaskDeck) Instead, we only need to create a new method to save a
list of cards in taskDeck and a new method to find the correct value
from the card.

As explained in https://github.com/ankidroid/Anki-Android/issues/5680
, this lead to a small difference. If a card type has a specified
front/back template, this would be used in the card browser. Unless
this lead to an empty string, in which case the usual template would
be used. This is not the case anymore, so that:
* we get closer to anki's feature.
* We don't actually have to modify libanki, which would be the natural
  place to put this feature's code.