Fix _checkDeckTree
I did forgot to add deck names in the name list. Which means that each
subdeck thought they don't have parents.
Parents were not created since _ensureParents founds the parents. And
it didn't made ankidroid really slower. However, it means that
duplicates were not found; and worse, it sent a lot of useless data to
the debugger, making it harder to read than what it was before