Correcting: onMark/onFlag raises nullPointException
When you click on mark/flag in the reviewer but the card is not yet
loaded, this used to lead to a nullPointException. This is now
corrected. I believe that it's okay if those button does nothing while
we are waiting for the next card.

Note that this leaves the related problem: when you review a card and
then click on "mark" on a slow phone, it's not clear which card is
marked; indeed, it depends on whether the value of mCurrentCard did
change or not.